requirements.txt
absl-py==1.2.0
antigravity==0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
bcc==0.18.0
beautifulsoup4==4.11.1
black==22.10.0
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
cachetools==5.2.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
chex==0.1.5
chrome-gnome-shell==0.0.0
click==8.0.3
cloudpickle==1.3.0
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.5.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.32
dbus-python==1.2.18
debugpy==1.6.2
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
dill==0.3.5.1
distro==1.7.0
dm-control==1.0.5
dm-env==1.5
dm-haiku==0.0.8
dm-tree==0.1.7
docopt==0.6.2
entrypoints==0.4
etils==0.8.0
execnet==1.9.0
executing==0.8.3
fasteners==0.17.3
fastjsonschema==2.16.1
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==2.1.3
flatbuffers==2.0.7
fonttools==4.34.4
future==0.18.2
gast==0.4.0
glfw==2.5.4
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.47.0
gym==0.17.2
gym-minigrid==1.0.3
gym-notices==0.0.7
h5py==3.7.0
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
imageio==2.21.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.6.4
importlib-resources==5.9.0
iniconfig==1.1.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jax==0.3.17
jaxlib==0.3.15
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
jmp==0.0.2
joblib==1.1.0
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.4
labmaze==1.0.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.6
lmdb==1.3.0
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==8.10.0
moviepy==1.0.3
mujoco==2.2.2
mujoco-py==2.1.2.14
mypy-extensions==0.4.3
natsort==8.0.2
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netaddr==0.8.0
netifaces==0.11.0
networkx==2.8.5
notebook==6.4.12
numpy==1.23.1
oauthlib==3.2.0
olefile==0.46
OpenNMT-py==2.2.0
opt-einsum==3.3.0
optax==0.1.3
opustools-pkg==0.0.52
packaging==21.3
pandadoc==0.1.0
pandas==1.4.3
pandoc==2.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
platformdirs==2.5.4
pluggy==1.0.0
plumbum==1.7.2
ply==3.11
pop-transition==1.1.2
portalocker==2.5.1
proglog==0.1.10
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.19.5
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydbus==0.6.0
pyglet==1.5.0
Pygments==2.12.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyonmttok==1.31.0
PyOpenGL==3.1.6
pyparsing==2.4.7
pyRFC3339==1.1
pyro-api==0.1.2
pyro-ppl==1.8.1
pyrsistent==0.18.1
pytest==7.1.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
PyWavelets==1.3.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.2.0
regex==2022.7.25
repolib==2.1.1
repoman==1.4.0
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
sacrebleu==2.2.0
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
SecretStorage==3.3.1
Send2Trash==1.8.0
sentencepiece==0.1.96
sessioninstaller==0.0.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
stack-data==0.3.0
systemd-python==234
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.10.0
tensorflow-datasets==4.6.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-metadata==1.10.0
termcolor==2.0.1
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.8.8
tinycss2==1.1.1
tinydb==4.7.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.0
torch-tb-profiler==0.4.0
torchmetrics==0.9.3
torchtext==0.5.0
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.6
waitress==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.1
wrapt==1.14.1
wxPython==4.0.7
xdg==5
xgboost==1.6.2
xkit==0.0.0
yarg==0.1.9
zipp==1.0.0
